% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_ratio.R
\name{abun_ratio}
\alias{abun_ratio}
\title{abun_ratio}
\usage{
abun_ratio(dat, group, ctrl = "ctrl_med")
}
\arguments{
\item{dat}{a data frame}

\item{group}{numerator}

\item{ctrl}{demoninator}
}
\description{
Abundance ratio function
}
\examples{
abun_ratio(df, "group1_med", "ctrl_med")
}
\keyword{group}
