% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_isos.R
\name{all_isos}
\alias{all_isos}
\title{Isotopic Distribution for all Pools}
\usage{
all_isos(dat, pool_cols, peptide)
}
\arguments{
\item{dat}{the model data frame}

\item{pool_cols}{all the pools in the model}

\item{peptide}{character string of the peptide}
}
\description{
Creates a vector containing the abundance ratio for each peptide
}
\examples{
all_isos(mod, pools, peptide)
}
\keyword{isos,}
\keyword{mega}
\keyword{model}
