% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_group.R
\name{as_group}
\alias{as_group}
\title{as_group}
\usage{
as_group(dat, col, FUN = median)
}
\arguments{
\item{dat}{data frame with peptide level data as rows and samples in columns}

\item{col}{a numeric vector containing the column numbers of the group
that the function will be performed on}
}
\description{
This function allows you to express your love of cats.
}
\examples{
as_group(df, c(4:6))
}
\keyword{group}
