% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_group.R
\name{by_group}
\alias{by_group}
\title{by_group}
\usage{
by_group(dat, col, FUN = median)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in columns}

\item{col}{a numeric vector containing the column numbers of the group
that the function will be performed on}
}
\description{
This function groups together
}
\examples{
by_group(df, c(4:6))
}
\keyword{group}
