% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_protein.R
\name{by_protein}
\alias{by_protein}
\title{FUN by protein}
\usage{
by_protein(dat, groups, FUN = mean)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in
columns}

\item{groups}{a vector of character strings or}

\item{FUN}{function argument for tapply() defaults to mean}
}
\description{
Takes a data frame and performs a function on each protein group within
"Master.Protein.Accessions" for specified groups of samples
}
\examples{
med_groups <- c("ctrl_med", "group1_med")
protein <- by_protein(df, med_groups)
}
\keyword{by}
\keyword{group}
\keyword{protein}
