% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iso.R
\name{combine_iso}
\alias{combine_iso}
\title{Sums Isotopes from Different Pools}
\usage{
combine_iso(dat)
}
\arguments{
\item{dat}{a data frame with multiple pools of individual isotopes to be summed}
}
\description{
Combines by summation each isotope from different pools
}
\examples{
mod <- cbind(mod, combine_iso(mod)) # as used in mega_model()
}
\keyword{isotope,}
\keyword{sum}
\keyword{total}
