% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_df.R
\name{count_df}
\alias{count_df}
\title{count_df}
\usage{
count_df(dat, col, dup = 1)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in columns}

\item{col}{a numeric vector containing the column numbers of the group to be
counted}
}
\description{
Function to count degrees of freedom in a group, defaults without duplicates
}
\examples{
count_df(df, c(4:13), dup = 3)
}
\keyword{degrees}
\keyword{freedom}
\keyword{of}
