% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_syn.R
\name{deg_syn}
\alias{deg_syn}
\title{Degradation and Synthesis of a Peptide Over Time}
\usage{
deg_syn(dat, deg, syn, initial, D3_pep, D3_i)
}
\arguments{
\item{dat}{a data frame containing columns with median peptide abundances}

\item{deg}{degradation constant (k)}

\item{syn}{synthesis rate (amount/time)}

\item{initial}{abundance at T0}

\item{D3_pep}{numeric vector containing the result from iso_dist}

\item{D3_i}{specific isotope (\emph{i.e.} 0 for "M+0", 1 for "M+1" \emph{etc.})}
}
\description{
How the pools change over time
}
\examples{
deg_syn(dat, f_deg, f_syn, f_initial, D3_pep,
        D3_i = as.numeric(gsub("^D3_([0-9]+)_.*", "\\\\1", pool)))
}
\keyword{degradation,}
\keyword{synthesis}
