% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_to_comp.R
\name{gene_to_comp}
\alias{gene_to_comp}
\title{Groups proteins by gene symbols into compartments or sub-compartments}
\usage{
gene_to_comp(dat, comp_dat, level = "compartment")
}
\arguments{
\item{dat}{a data frame with gene symbols}

\item{level}{a string containing the level to be sorted}

\item{gene_dat}{a reference data frame with gene symbols and the assigned
compartments and sub-compartments within the cell}
}
\description{
Takes data frame with a "gene" column and adds a column using a second data
frame as a reference to convert between "Gene" column and a compartment
column
}
\examples{
df$sub_compartment <- gene_to_comp(df, comp_df, level = "sub_compartment")
}
\keyword{compartments}
\keyword{sub-cellular}
