% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_mpa_to_gene.R
\name{indiv_mpa_to_gene}
\alias{indiv_mpa_to_gene}
\title{Converts Accession to Gene Name for Individual Proteins}
\usage{
indiv_mpa_to_gene(Acc_pro, gene_dat)
}
\arguments{
\item{Acc_pro}{accession number to be converted to gene}

\item{gene_dat}{reference data frame with Master.Protein.Accessions and gene columns}
}
\description{
Looks up a accession in a gene_df and finds the corresponding gene name
}
\examples{
gene <- indiv_mpa_to_gene(pro, protein)
}
\keyword{accession,}
\keyword{gene}
\keyword{master}
\keyword{protein}
