% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_dist.R
\name{iso_dist}
\alias{iso_dist}
\title{Distribution of D3-Leu peptides}
\usage{
iso_dist(pep, p = 0.5)
}
\arguments{
\item{pep}{peptide, must contain at least one "L"}

\item{p}{the percent of D3-labeled Leucine in the labeled media, defaults to 0.5}
}
\description{
Calculates the distribution of newly synthesized D3-Leu peptides
}
\section{Warning}{

p cannot equal 1, if the percent label is 100%, then write p = 0.9999
}

\examples{
iso_dist("SAMPLLER", p = 0.99)
}
\keyword{D3}
\keyword{Leucine,}
\keyword{labeling,}
\keyword{peptide,}
\keyword{synthesis}
