% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ratio_calc.R
\name{iso_ratio_calc}
\alias{iso_ratio_calc}
\title{"Corrected" Isotope Ratio}
\usage{
iso_ratio_calc(dat)
}
\arguments{
\item{dat}{a data frame containing the grouping variables, the isotopes and the T0 ratio}
}
\description{
Creates a vector containing the abundance ratio for each peptide
}
\examples{
df$ratio <- iso_ratio_calc(df)
}
\keyword{corrected}
\keyword{isotope}
\keyword{ratio}
