% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_yf.R
\name{k_yf}
\alias{k_yf}
\title{K constant and Plateau from Non-Linear Fit}
\usage{
k_yf(nls_df, leg)
}
\arguments{
\item{nls_df}{non-linear fit data frame}

\item{leg}{which leg group}
}
\description{
Creates an object with the k and yf values
}
\examples{
  df <- dat \%>\%
group_by(leg) \%>\%
  do(fit = nls(ratio ~ SSasymp(week, yf, y0, log_k), data = .)) \%>\%
  tidy(fit) \%>\%
  select(leg, term, estimate) \%>\%
  spread(term, estimate) \%>\%
  mutate(k = exp(log_k))

L_k_yf <- k_yf(df, "L")
}
\keyword{fit,}
\keyword{k,}
\keyword{non-linear}
\keyword{plateau}
\keyword{plot,}
