% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_fit.R
\name{lin_fit}
\alias{lin_fit}
\title{Linear fit result}
\usage{
lin_fit(dat)
}
\arguments{
\item{dat}{a data frame containing abundance versus week}
}
\description{
Creates a data frame with intercept, slope, r^2, and p-value columns
}
\examples{
df_fit <- df \%>\%
  group_by(Master.Protein.Accessions, sex, leg) \%>\%
  do(lin_fit(.))
}
\keyword{fit,}
\keyword{intercept,}
\keyword{linear}
\keyword{r-squared}
\keyword{slope,}
