% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eqn.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Linear Regression Equation Object}
\usage{
lm_eqn(lm_object)
}
\arguments{
\item{lm_object}{linear fit object}
}
\description{
Creates an character vector with the y = mx + b equation, r-squared, and p-value, to be used for adding to plots
}
\examples{
df_l <- filter(dat, leg == "L")
lm_l <- lm(abundance ~ week, df_l)
eqn_l <- lm_eqn(lm_l)
}
\keyword{equation,}
\keyword{linear}
\keyword{p-value,}
\keyword{plot}
\keyword{r-squared,}
\keyword{regression,}
