% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pools.R
\name{make_pools}
\alias{make_pools}
\title{Make pools for D3_0_old, D3_0_new, D3_1, D3_2 \emph{etc.} Pools}
\usage{
make_pools(dat, pool_names, syn, deg_old, deg_new, t0_abun, D3_pep)
}
\arguments{
\item{dat}{data frame}

\item{pool_names}{vector containing the pool names for each pool of labeled Leucines}

\item{syn}{synthesis rate of new material}

\item{deg_old}{degredation constant for the "old" material present at T0}

\item{deg_new}{degredation constant for the newly synthesized material}

\item{t0_abun}{abundance at the start (T0)}

\item{D3_pep}{numeric vector containing the result from iso_dist}
}
\description{
Part of \strong{MEGA MODEL}
}
\examples{
pool_df <- make_pools(mod, pool_names, syn, deg_old, deg_new, t0_abun,
                      D3_pep_dist)
}
\keyword{degredation,}
\keyword{pools}
\keyword{synthesis,}
