% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega_model.R
\name{mega_model}
\alias{mega_model}
\title{MEGA MODEL}
\usage{
mega_model(peptide, deg_old, deg_new, syn, t0_abun, per_lab, time)
}
\arguments{
\item{peptide}{peptide containing Leucine}

\item{deg_old}{degradation rate constant (k) of the "old" material around at the beginning of the experiment}

\item{deg_new}{degradation rate constant (k) of newly synthesized material}

\item{syn}{synthesis rate (amount/time)}

\item{t0_abun}{initial abundance}

\item{per_lab}{percent of D3-Leu in the labeled growth media}

\item{time}{vector of time to be modeled (\emph{e.g.} 0:168)}
}
\description{
Creates a data frame with the change of individual isotopes over time
}
\examples{
# initial conditions
peptide <- "SAMPLLLLLLER"

# initial total abundance at T0
t0_abun <- 1000

# fraction of labeled Leucine in the media
per_lab <- 0.50

# rates
deg_old <- 0.0500
deg_new <- 0.0500
syn <- 50

# length of time
time <- 0:168

mod <- mega_model(peptide, deg_old, deg_new, syn, t0_abun, per_lab, time)
}
\keyword{degredation}
\keyword{mega-model,}
\keyword{peptide}
\keyword{synthesis,}
\keyword{turnover,}
