% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_to_gene.R
\name{mpa_to_gene}
\alias{mpa_to_gene}
\title{mpa_to_gene}
\usage{
mpa_to_gene(dat, gene_dat)
}
\arguments{
\item{dat}{a data frame with accession numbers}

\item{gene_dat}{a reference data frame with all accession numbers and gene symbols}
}
\description{
Convert Master.Protein.Accession to gene symbol
}
\examples{
mpa_to_gene(df, gene_df)
}
\keyword{gene}
\keyword{symbol}
