% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_iso.R
\name{norm_iso}
\alias{norm_iso}
\title{Normalize Isotopes to the Sum}
\usage{
norm_iso(dat)
}
\arguments{
\item{dat}{a data frame containing columns of all individual isotopes ("M_0" \emph{etc.}) and "Sum"}
}
\description{
Adds columns of normalized isotopes
}
\examples{
df <- df_tidy \%>\%
  spread(., "isotope", "abundance") \%>\%
  as.data.frame(.) \%>\%
  norm_iso(.)
}
\keyword{isotopes}
\keyword{normalize}
