% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_isos.R
\name{plot_all_isos}
\alias{plot_all_isos}
\title{Plot the Isotopic Distribution Changing Over Time}
\usage{
plot_all_isos(dat, g_title, g_subtitle, FUN = geom_jitter)
}
\arguments{
\item{dat}{a data frame containing isotope and abundance}

\item{g_title}{title for the plot}

\item{g_subtitle}{subtitle for the plot}

\item{FUN}{geom_jitter or geom_point (defaults to geom_jitter)}
}
\description{
Creates a vector containing the abundance ratio for each peptide
}
\examples{
df_g <- filter(df_iso, protein == "Hist2h4", peptide == "DAVTYTEHAK")
plot_all_isos(df_g, "PRO", "PEP")
}
\keyword{distribution}
\keyword{isotopic}
\keyword{plot,}
