% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pro.R
\name{plot_pro}
\alias{plot_pro}
\title{Plot proteins changing overtime}
\usage{
plot_pro(dat, g_title, FUN = geom_jitter, x_pos = 9, y1_pos = 0.83,
  y2_pos = 0.3)
}
\arguments{
\item{dat}{a data frame containing abundance and week columns}

\item{g_title}{title of the plot}

\item{FUN}{geom_jitter or geom_point}

\item{x_pos, y1_pos, y2_pos}{x and y coordinates of the linear eqn and r-squared objects}
}
\description{
Plots changing protein abundance over time for two different groups
}
\examples{
pros <- as.character(unique(df$Master.Protein.Accessions))

plot_list <- list()

for (pro in pros){
  gene <- indiv_mpa_to_gene(pro, protein)
  temp_df <- filter(df, Master.Protein.Accessions == pro)
  g <- plot_pro(temp_df, g_title = gene)
  plot_list[[pro]] <- g
}

pdf("all_plots.pdf", width = 10.75, height = 6)

for(pro in pros){
  print(plot_list[[pro]])
}
}
\keyword{abundance,}
\keyword{plot,}
\keyword{protein}
\keyword{time}
