% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_iso_dist.R
\name{pool_iso_dist}
\alias{pool_iso_dist}
\title{Isotopic Distribution of a Pool}
\usage{
pool_iso_dist(dat, pool, peptide)
}
\arguments{
\item{dat}{data frame with the pool in it}

\item{pool}{specific D3_0 \emph{etc.} pool}

\item{peptide}{peptide}
}
\description{
Multiplies a pool by the isotopic distribution
}
\examples{
temp <- pool_iso_dist(dat, pools, peptide) # as used in all_isos() function
}
\keyword{distribution}
\keyword{isotopic}
\keyword{pool,}
