% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome.R
\name{proteome}
\alias{proteome}
\title{Proteome}
\usage{
proteome(file_name, organism, wt_samps = "all", group = FALSE,
  group_names = NULL, top_pep = 3, min_pep = 3, norm = FALSE,
  norm_method = NULL, norm_pro = NULL, csv = TRUE)
}
\arguments{
\item{file_name}{file containing all peptides for all samples with the columns named Annotated.Sequence, Modifications, Master.Protein.Accessions.}

\item{organism}{"mouse" or "human"}

\item{wt_samps}{"all" if you want to use all samples to weight, otherwise use a pattern that is present in the column names of the samples you wish to use for the weighting}

\item{group}{if you want to}

\item{group_names}{indicate names of the groups (must be a string pattern present in the samples)}

\item{top_pep}{number of top ionizers used}

\item{min_pep}{minumum number of peptides found in a protein to be incuded in the analysis}

\item{norm}{set to TRUE if normalization is wanted}

\item{norm_method}{indicate normalization method: "protein" or "sum_total"}

\item{norm_pro}{accession of protein used for}

\item{csv}{if result file should be written in working directory}
}
\description{
Creates a data frame with an average abundance of the top ionizers for each protein
}
\examples{
proteome("WT vs KO all pep.csv", group_names = c("WT", "KO"),
         organism = "mouse", group = TRUE, norm = TRUE,
         norm_method = "protein", norm_pro = "B2RTM0", csv = FALSE)
}
\keyword{analysis}
\keyword{proteome}
\keyword{whole}
