% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_ttest.R
\name{pval_ttest}
\alias{pval_ttest}
\title{t-test function between groups}
\usage{
pval_ttest(dat, group, ctrl)
}
\arguments{
\item{dat}{a data frame stacked}
}
\description{
This function groups together
}
\examples{
pval_ttest(df, "group1", "group2")
}
\keyword{ttest}
