% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_sd.R
\name{ratio_sd}
\alias{ratio_sd}
\title{ratio_sd}
\usage{
ratio_sd(dat, ratio, sd1, med1, sd2, med2)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in columns}
}
\description{
Taylor Expansion to get the sd of the ratio of two means
}
\examples{
ratio_sd(df, "ratio", "group1_sd", "group1_med", "ctrl_sd", "ctrl_med")
}
\keyword{group}
