% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outliers.R
\name{rm_outliers}
\alias{rm_outliers}
\title{rm_outliers}
\usage{
rm_outliers(dat, pro_df, ratio)
}
\arguments{
\item{dat}{a data frame}

\item{pro_df}{a data fram containing average ratio by protein}

\item{ratio}{specific ratio to remove outliers by}
}
\description{
Remove outliers in peptide data
}
\examples{
rm_outliers(df, pro_df, "ratio")
}
\keyword{group}
