% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_x_df.R
\name{square_x_df}
\alias{square_x_df}
\title{square_x_df}
\usage{
square_x_df(dat, sd, df)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in columns}
}
\description{
Standard deviation of grouped relative abundance using top ionizers
}
\examples{
square_x_df(df,"group1_sd", "group1_df")
}
\keyword{calc}
\keyword{sd}
