% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_ratio.R
\name{abun_ratio}
\alias{abun_ratio}
\title{Abundance ratio}
\usage{
abun_ratio(dat, group, ctrl = "ctrl_med")
}
\arguments{
\item{dat}{a data frame containing columns with median peptide abundances}

\item{group}{column string or number (numerator of ratio)}

\item{ctrl}{column string or number (demoninator of ratio)
defaults to "ctrl_med"}
}
\description{
Creates a vector containing the abundance ratio for each peptide
}
\examples{
df$ratio <- abun_ratio(df, "group1_med")
}
\keyword{ratio}
