% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_group.R
\name{by_group}
\alias{by_group}
\title{Peptide average or standard deviation}
\usage{
by_group(dat, col, FUN = median)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in
columns}

\item{col}{a numeric vector containing the column numbers of samples within
the group}
}
\description{
Peptide level average or sd of a group of samples
}
\examples{
# selects the columns containing a string that identifies the group
group1 <- grep("Sample_norm", colnames(data))
group1_sd <- by_group(df, group1, FUN = sd)
}
\keyword{group}
\keyword{median}
\keyword{sd}
