% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_group.R
\name{by_protein}
\alias{by_protein}
\title{by_protein}
\usage{
by_protein(dat, groups, FUN = mean)
}
\arguments{
\item{dat}{a data frame with peptide level data as rows and samples in columns}

\item{groups}{a numeric vector containing the column numbers of the group
that the function will be performed on}
}
\description{
Groups by protein
}
\examples{
protein_group(df, c("ctrl_med", "group1_med"))
}
\keyword{by}
\keyword{group}
\keyword{protein}
