% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outliers.R
\name{rm_outliers}
\alias{rm_outliers}
\title{Remove outliers in peptide data}
\usage{
rm_outliers(dat, pro_df, ratio, mult = 2)
}
\arguments{
\item{dat}{a data frame}

\item{pro_df}{a data fram containing average ratio of each protein}

\item{ratio}{a character string of the column name of the ratio used to find
outliers}

\item{mult}{multiplier of the sd used to determine the min and max allowable
ratio for each peptide, defaults to 2}
}
\description{
Creates a new data frame with the outliers from original data frame removed.
Outliers are defined as ratios
}
\examples{
data_rm_out <- rm_outliers(df, pro_df, "ratio")
}
\keyword{outliers}
\keyword{remove}
